requirements.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Database dependencies
pymysql>=1.1.0
cryptography>=41.0.0

# Image processing
Pillow>=10.0.0

# Data handling
openpyxl>=3.1.0
xlrd>=2.0.0

# Google Drive API dependencies
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.176.0
google-auth==2.40.3

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Optional: For better performance
pyarrow>=14.0.0 